refine test in repartition.rs & coalesce_batches.rs
